refactor!(crates/http): extract git route handlers to `git` submodule
The functions were defined in the top-level scope of the handlers
module. Now they are in `thoenix_http::handlers::git` to match the `tf`
module.
